backend/requirements.txt
Django>=5.0,<7.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=24.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
Pillow>=10.0.0
cloudinary>=1.41.0
django-cloudinary-storage>=0.3.0
sentry-sdk[django]>=2.14.0
